[feat] separate node kind and node pattern
We need to separate Pattern into two separate struct.
The reason is that Replacer trait cannot be implemented on Pattern enum.
If we just use str as replacer, this will parse the str every time a replacement happens.
Instead, we can use node Pattern to store the replacement AST.
Note, Node already implements Replacer trait and it does not replace metavar. So we cannot use
Node/Root to store the Replacer.