fix: use get_range instead as_slice to handle byte_offset and underlying arry size
We used to expose underlying binary array of the Source. However, in JavaScript.
The start_pos is in byte offset.  But underlying binary array can be in u16.
This pull request changes the interface of source by replacing as_slice with get_range.
So source can handle this discrepancy gracefully.
fix #400