fix: loading of custom languages
* fix: loading of custom languages

The argument order here was swapped. This caused custom languages to not
be registered for commands like `new rule`. This patch passes arguments
in correct order.

* fix: stringification of `SgLang`

The impls of `Display` and `Debug` for `SgLang` seem to have been
swapped. This caused custom languages to be listed in their debug
representation e.g., during `rule new`. With this patch `Display` and
`Debug` are implemented like one would naively expect.