feat: better metavar detection
BREAKING CHANGE: metavar in pattern is more intuitive now.
Previously we only treat leaf node and error node as metavar.
This commit will detect metavar as long as the text in the node
matches metavar regex pattern. This results in a more intuitive
pattern syntax. However, this introduces 5% perf overhead and
is a breaking change. Breaking change is ok since the matching
is better now.