feat(cli): Add max-count flag to limit matches
Added a new feature to limit the number of matches processed, similar to
grep's `-m/--max-count` flag.

- Updated `InputArgs` struct to include `max_count`.
- Modified `run.rs`, `scan.rs` files to pass `max_count` into various worker implementations.

This update enhances the tool's capabilities for match processing,
allowing users to limit the output to a specified number of entries.
Specifically, this arose when writing go code to turn a certain subset
of `=` assignments to `:=` declarations, which we'd only want to apply
once.

Previously this was handled in the rule config by some extra (relatively
complex) specification resembling:
```
subsequent-match:
  pattern: $LOG_VAR = $_
  ... (extra context)
```
and then including a `rules.all.not` to ensure `subsequent-match` was
false. This would be great to support (and follow the config used by
grep).

Also mentioned in https://github.com/ast-grep/ast-grep/issues/1296.