refactor: remove potential dynamic library footgun
This commit eliminates the need to continually keep a reference to the
dynamic language library to prevent it from getting cleaned up by
`libloading`, leaving us with dangling pointers. Instead, we can use
`std::mem::forget()` to tell the compiler to forget to call its
destructor when dropped. This approach allows us to not worry about
having to keep references to the library in order to prevent null
pointers, and it is also the approach taken by the `helix` editor so it
seems like a trusty solution.